Handle "d." money strings.
BigNumber does not handle a string like "1." so adding support to the
Money constructure to look for "d." format and convert it to "d.0"
before processing. This is done in Money to have a general solution to
ensure such strings are handled. Further restrictions could be added to
validators or elsewhere to restrict UI usage if needed.